Move helper to utils
